README: CLI supports service env vars now
https://github.com/balena-io/balena-cli/pull/1529

This PR merged support for setting env vars on microservices
via the CLI so we can remove this text about it not being supported.

We could also improve the doc by adding an example of setting a
service env var via the CLI but this PR is just to remove the
unsupported reference.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>